Phase 2 commit 2d: hoist GlobalMemory, VirtualMemory, Sensors native-free code
Move the public API surface, memoized suppliers, and command-line bits
of FreeBsdGlobalMemory, FreeBsdVirtualMemory, and FreeBsdSensors onto
the oshi-common abstracts; the JNA concretes shrink to just the
sysctl-bound query hooks.

FreeBsdGlobalMemory abstract:
- Owns the available/total/pageSize/vm suppliers and getter methods.
- queryPageSize() lives here as a static command-line helper (sysconf
  PAGESIZE; sysctl hw.pagesize doesn't work on FreeBSD 13).
- queryVmStats(), queryPhysMem(), createVirtualMemory() are protected
  abstract hooks.

FreeBsdVirtualMemory abstract:
- Owns swap suppliers, getters, and the swapinfo -k parsing in
  querySwapUsed().
- Constructor takes a FreeBsdGlobalMemory so the abstract can compute
  getVirtualMax/InUse from getTotal()/getAvailable().
- querySwapTotal(), queryPagesIn(), queryPagesOut() are protected
  abstract hooks.

FreeBsdSensors abstract:
- Inherits no FreeBSD-supported sources for fan speed or CPU voltage;
  those become concrete no-ops here so JNA/FFM don't both redeclare.
- queryCpuTemperature() remains abstract — coretemp sysctl per CPU.

JNA concretes (FreeBsdGlobalMemoryJNA, FreeBsdVirtualMemoryJNA,
FreeBsdSensorsJNA) now contain only the BsdSysctlUtil and FreeBsdLibc
sysctl reads. FreeBsdVirtualMemoryJNA constructor accepts the abstract
FreeBsdGlobalMemory type since all calls it makes (getTotal,
getAvailable) live on the abstract.
