+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/freebsd/FreeBsdSensors.java
@@ -7,4 +7,16 @@ package oshi.hardware.common.platform.unix.freebsd;
 import oshi.hardware.common.AbstractSensors;
 
 public abstract class FreeBsdSensors extends AbstractSensors {
+
+    @Override
+    public int[] queryFanSpeeds() {
+        // Nothing known on FreeBSD for this.
+        return new int[0];
+    }
+
+    @Override
+    public double queryCpuVoltage() {
+        // Nothing known on FreeBSD for this.
+        return 0d;
+    }
 }
