+++ oshi-core/src/main/java/oshi/hardware/platform/unix/freebsd/FreeBsdGlobalMemoryJNA.java
@@ -4,69 +4,33 @@
  */
 package oshi.hardware.platform.unix.freebsd;
 
-import static oshi.util.Memoizer.defaultExpiration;
-import static oshi.util.Memoizer.memoize;
-
-import java.util.function.Supplier;
-
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.hardware.VirtualMemory;
 import oshi.hardware.common.platform.unix.freebsd.FreeBsdGlobalMemory;
-import oshi.util.ExecutingCommand;
-import oshi.util.ParseUtil;
 import oshi.util.platform.unix.freebsd.BsdSysctlUtil;
 
 /**
- * Memory obtained by sysctl vm.stats
+ * JNA-backed FreeBSD global memory. Native reads use {@link BsdSysctlUtil}; shared parsing and the public API live on
+ * {@link FreeBsdGlobalMemory}.
  */
 @ThreadSafe
 public class FreeBsdGlobalMemoryJNA extends FreeBsdGlobalMemory {
 
-    private final Supplier<Long> available = memoize(this::queryVmStats, defaultExpiration());
-
-    private final Supplier<Long> total = memoize(FreeBsdGlobalMemoryJNA::queryPhysMem);
-
-    private final Supplier<Long> pageSize = memoize(FreeBsdGlobalMemoryJNA::queryPageSize);
-
-    private final Supplier<VirtualMemory> vm = memoize(this::createVirtualMemory);
-
     @Override
-    public long getAvailable() {
-        return available.get();
-    }
-
-    @Override
-    public long getTotal() {
-        return total.get();
-    }
-
-    @Override
-    public long getPageSize() {
-        return pageSize.get();
-    }
-
-    @Override
-    public VirtualMemory getVirtualMemory() {
-        return vm.get();
-    }
-
-    private long queryVmStats() {
+    protected long queryVmStats() {
         // cached removed in FreeBSD 12 but was always set to 0
         int inactive = BsdSysctlUtil.sysctl("vm.stats.vm.v_inactive_count", 0);
         int free = BsdSysctlUtil.sysctl("vm.stats.vm.v_free_count", 0);
         return (inactive + free) * getPageSize();
     }
 
-    private static long queryPhysMem() {
+    @Override
+    protected long queryPhysMem() {
         return BsdSysctlUtil.sysctl("hw.physmem", 0L);
     }
 
-    private static long queryPageSize() {
-        // sysctl hw.pagesize doesn't work on FreeBSD 13
-        return ParseUtil.parseLongOrDefault(ExecutingCommand.getFirstAnswer("sysconf PAGESIZE"), 4096L);
-    }
-
-    private VirtualMemory createVirtualMemory() {
+    @Override
+    protected VirtualMemory createVirtualMemory() {
         return new FreeBsdVirtualMemoryJNA(this);
     }
 }
