+++ oshi-core/src/main/java/oshi/hardware/platform/unix/freebsd/FreeBsdVirtualMemoryJNA.java
@@ -4,85 +4,34 @@
  */
 package oshi.hardware.platform.unix.freebsd;
 
-import static oshi.util.Memoizer.defaultExpiration;
-import static oshi.util.Memoizer.memoize;
-
-import java.util.function.Supplier;
-
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.hardware.common.platform.unix.freebsd.FreeBsdGlobalMemory;
 import oshi.hardware.common.platform.unix.freebsd.FreeBsdVirtualMemory;
-import oshi.util.ExecutingCommand;
-import oshi.util.ParseUtil;
 import oshi.util.platform.unix.freebsd.BsdSysctlUtil;
 
 /**
- * Memory obtained by swapinfo
+ * JNA-backed FreeBSD virtual memory. Native sysctl reads use {@link BsdSysctlUtil}; the public API and the swapinfo
+ * command-line parsing live on {@link FreeBsdVirtualMemory}.
  */
 @ThreadSafe
 final class FreeBsdVirtualMemoryJNA extends FreeBsdVirtualMemory {
 
-    private final FreeBsdGlobalMemoryJNA global;
-
-    private final Supplier<Long> used = memoize(FreeBsdVirtualMemoryJNA::querySwapUsed, defaultExpiration());
-
-    private final Supplier<Long> total = memoize(FreeBsdVirtualMemoryJNA::querySwapTotal, defaultExpiration());
-
-    private final Supplier<Long> pagesIn = memoize(FreeBsdVirtualMemoryJNA::queryPagesIn, defaultExpiration());
-
-    private final Supplier<Long> pagesOut = memoize(FreeBsdVirtualMemoryJNA::queryPagesOut, defaultExpiration());
-
-    FreeBsdVirtualMemoryJNA(FreeBsdGlobalMemoryJNA freeBsdGlobalMemory) {
-        this.global = freeBsdGlobalMemory;
+    FreeBsdVirtualMemoryJNA(FreeBsdGlobalMemory global) {
+        super(global);
     }
 
     @Override
-    public long getSwapUsed() {
-        return used.get();
-    }
-
-    @Override
-    public long getSwapTotal() {
-        return total.get();
-    }
-
-    @Override
-    public long getVirtualMax() {
-        return this.global.getTotal() + getSwapTotal();
-    }
-
-    @Override
-    public long getVirtualInUse() {
-        return this.global.getTotal() - this.global.getAvailable() + getSwapUsed();
-    }
-
-    @Override
-    public long getSwapPagesIn() {
-        return pagesIn.get();
-    }
-
-    @Override
-    public long getSwapPagesOut() {
-        return pagesOut.get();
-    }
-
-    private static long querySwapUsed() {
-        String swapInfo = ExecutingCommand.getAnswerAt("swapinfo -k", 1);
-        String[] split = ParseUtil.whitespaces.split(swapInfo);
-        if (split.length < 5) {
-            return 0L;
-        }
-        return ParseUtil.parseLongOrDefault(split[2], 0L) << 10;
-    }
-
-    private static long querySwapTotal() {
+    protected long querySwapTotal() {
         return BsdSysctlUtil.sysctl("vm.swap_total", 0L);
     }
 
-    private static long queryPagesIn() {
+    @Override
+    protected long queryPagesIn() {
         return BsdSysctlUtil.sysctl("vm.stats.vm.v_swappgsin", 0L);
     }
 
-    private static long queryPagesOut() {
+    @Override
+    protected long queryPagesOut() {
         return BsdSysctlUtil.sysctl("vm.stats.vm.v_swappgsout", 0L);
     }
 }
